Resolved issues with Widget::draw implementation and on_event return type raised after the iced_native upgrade to 0.3. Implemented a softer highlight color with opacity set to 0.5

+++ crates/widgets/src/style/table_row.rs
@@ -6,8 +6,8 @@ use iced_core::{Background, Color};
 pub struct Style {
     pub text_color: Option<Color>,
     pub background: Option<Background>,
-    pub border_radius: u16,
-    pub border_width: u16,
+    pub border_radius: f32,
+    pub border_width: f32,
     pub border_color: Color,
 }
 
@@ -26,8 +26,8 @@ impl StyleSheet for Default {
         Style {
             text_color: None,
             background: None,
-            border_radius: 0,
-            border_width: 0,
+            border_radius: 0.0,
+            border_width: 0.0,
             border_color: Color::TRANSPARENT,
         }
     }
